+++ cmp/viewmanager/DataAccess.ts
@@ -34,7 +34,7 @@ export class DataAccess<T> {
             const ret = await XH.fetchJson({
                 url: 'xhView/allData',
                 params: {type, viewInstance: instance},
-                span: {name: 'getAllViews', source: 'hoist', caller: this}
+                span: {name: 'xh.getAllViews', caller: this}
             });
             return {
                 views: ret.views.map(v => new ViewInfo(v, this.model)),
@@ -56,7 +56,7 @@ export class DataAccess<T> {
             const raw = await XH.fetchJson({
                 url: 'xhView/get',
                 params: {token},
-                span: {name: 'getView', source: 'hoist', caller: this}
+                span: {name: 'xh.getView', caller: this}
             });
             return View.fromBlob(raw, model);
         } catch (e) {
@@ -71,7 +71,7 @@ export class DataAccess<T> {
             const raw = await XH.postJson({
                 url: 'xhView/create',
                 body: {type: model.type, ...spec},
-                span: {name: 'createView', source: 'hoist', caller: this}
+                span: {name: 'xh.createView', caller: this}
             });
             return View.fromBlob(raw, model);
         } catch (e) {
@@ -87,7 +87,7 @@ export class DataAccess<T> {
                 url: 'xhView/updateInfo',
                 params: {token: view.token},
                 body: updates,
-                span: {name: 'updateViewInfo', source: 'hoist', caller: this}
+                span: {name: 'xh.updateViewInfo', caller: this}
             });
             return View.fromBlob(raw, this.model);
         } catch (e) {
@@ -103,7 +103,7 @@ export class DataAccess<T> {
                 url: 'xhView/updateValue',
                 params: {token: view.token},
                 body: value,
-                span: {name: 'updateViewValue', source: 'hoist', caller: this}
+                span: {name: 'xh.updateViewValue', caller: this}
             });
             return View.fromBlob(raw, this.model);
         } catch (e) {
@@ -120,7 +120,7 @@ export class DataAccess<T> {
             await XH.postJson({
                 url: 'xhView/delete',
                 params: {tokens: map(views, 'token').join(',')},
-                span: {name: 'deleteViews', source: 'hoist', caller: this}
+                span: {name: 'xh.deleteViews', caller: this}
             });
         } catch (e) {
             throw XH.exception({
@@ -139,7 +139,7 @@ export class DataAccess<T> {
             url: 'xhView/updateState',
             params: {type, viewInstance: instance},
             body: update,
-            span: {name: 'updateViewState', source: 'hoist', caller: this}
+            span: {name: 'xh.updateViewState', caller: this}
         });
     }