+++ svc/ChangelogService.ts
@@ -9,6 +9,7 @@ import {HoistService, XH} from '@xh/hoist/core';
 import jsonFromMarkdown from '@xh/app-changelog.json';
 import {action, makeObservable, observable} from '@xh/hoist/mobx';
 import {checkMinVersion} from '@xh/hoist/utils/js';
+import {Span} from '@xh/hoist/utils/telemetry';
 import {isEmpty, forOwn, includes} from 'lodash';
 
 /**
@@ -82,7 +83,7 @@ export class ChangelogService extends HoistService {
         makeObservable(this);
     }
 
-    override async initAsync() {
+    override async initAsync(span: Span) {
         this.changelog = !isEmpty(jsonFromMarkdown?.versions)
             ? jsonFromMarkdown
             : {title: null, versions: []};