+++ svc/ConfigService.ts
@@ -6,6 +6,7 @@
  */
 import {HoistService, XH} from '@xh/hoist/core';
 import {deepFreeze, throwIf} from '@xh/hoist/utils/js';
+import {Span} from '@xh/hoist/utils/telemetry';
 import {keys} from 'lodash';
 
 /**
@@ -28,12 +29,10 @@ export class ConfigService extends HoistService {
 
     private _data = {};
 
-    override async initAsync() {
-        this._data = await XH.fetchJson({
-            url: 'xh/getConfig',
-            span: {name: 'getConfig', source: 'hoist', caller: this}
-        });
+    override async initAsync(span: Span) {
+        this._data = await XH.fetchJson({url: 'xh/getConfig', span});
         deepFreeze(this._data);
+        XH.traceService.noteConfigAvailable();
     }
 
     /**